+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Kirby Index Block
 
+## 1.0.3
+
+- **BREAKING:** Replaced wrong usage of ARIA-Attribute `aria-current` with a CSS-Class.
+  *Requires adjustments in your CSS-Rules.*
+  - New: Option `paginationCurrentPageCssClass` (default: `active`), field and HTML-Output applied to pagination `li`
+  - Change: Moved the ARIA-Attribute/Value `aria-current="page"` to the pagination link-tag `a` (instead of `li`)
+- README.MD
+  - New: Git-Submodule, composer installation
+
 ## 1.0.2
 
 - Added french translation (thanks [@joachimesque](https://github.com/joachimesque))